requirements.txt
mss
winrt-Windows.Media.Ocr
winrt-Windows.Graphics.Imaging
winrt-Windows.Storage.Streams
winrt-Windows.Foundation.Collections
winrt-Windows.Foundation
winrt-Windows.Globalization
requests
pillow
keyboard
urllib3
pygetwindow
